Eliminate Win32 type leakage from src/core (Phase 4)
Replaced all Win32-specific type declarations in src/core with platform-
agnostic types, completing the Platform Abstraction Layer (PAL) pattern.

## Platform Type Additions (types.h)
- ThreadHandle: Abstract thread handle
- MutexHandle: Abstract mutex handle
- EventHandle: Abstract event handle
- ModuleHandle: Abstract module handle
- OverlappedHandle: Abstract overlapped I/O handle

## Engine Subsystem (5 files)
engine.h:
- HANDLE → ThreadHandle/MutexHandle/EventHandle
- HGLOBAL → void*
- WPARAM_T/LPARAM_T → WParamT/LParamT (naming convention)

engine.cpp:
- Wrapped Windows sync code in #ifdef _WIN32

engine_lifecycle.cpp:
- HANDLE → ThreadHandle/MutexHandle/EventHandle
- unsigned int → uint32_t
- Isolated Windows CreateThread/CreateMutex/CreateEvent

engine_window.cpp:
- ULONG_PTR → uintptr_t
- LONG → int32_t

engine_focus.cpp:
- DWORD → uint32_t

## Window Subsystem (3 files)
target.cpp:
- HWND member vars → WindowHandle
- WPARAM/LPARAM → uintptr_t/intptr_t
- Kept Windows callbacks with proper casts

layoutmanager.cpp:
- HWND → WindowHandle
- WPARAM/LPARAM → uintptr_t/intptr_t
- Replaced GET_X_LPARAM/GET_Y_LPARAM with manual extraction

focus.cpp/h:
- ATOM → uint16_t (Register_focus return type)

## Settings (2 files)
setting.cpp:
- DWORD → uint32_t

setting.h:
- LONG → int32_t

## Results
- Build: ✅ Successful
- Win32 types in src/core: Minimal (only in Windows-specific blocks)
- Architecture: Clean platform abstraction maintained
- Zero windows.h includes in headers maintained

Remaining Win32 types are acceptable (Windows callbacks, #ifdef blocks).

+++ src/core/settings/setting.h
@@ -30,7 +30,7 @@ public:
     bool m_sts4mayu;                ///
     bool m_cts4mayu;                ///
     bool m_mouseEvent;                ///
-    LONG m_dragThreshold;            ///
+    int32_t m_dragThreshold;            ///
     unsigned int m_oneShotRepeatableDelay;    ///
 
 public:
